Changed color model. Fixed save-load wrt color.
Block colors are saved in hue/value pairs rather
than rgb. All auxilliary colors are generated from
the hv pair. The file format which used to save
9 rgb numbers now saves 2 floats, h and v.

I fixed a bug which crashed the program if you
load the old file format with a line of 9 numbers
rather than 2. This is probably one of many such
file related bugs.