Added input selector. Broke track settings.
You can press the left-most slot to select which
track gets keyboard and midi input. This will be
used in the future when recording is added.

Track settings UI got changed and bank was added.
Currently the four widgets do not display a meaningful
label, and do not function when manipulated. A
reason for this change was to allow changing the
track initial program setting while seeing the effects
of the 'auto track name' (GM patch names) appear in
the name input field. It also looks better (or will
look better) than the old style label and arrows
widget.

I also implicitly limited, in the gui, the number
of tracks to 16. This limit was already in place
but I planned to allow a way to add more tracks.
Nothing about the internals or file format limits
the number of tracks to 16, in principle. At the
moment this is cleaner. In the future a GUI mechanism
to change the number of tracks may be added.