Added auto track naming option.
If the option is activated then changing the program track
setting will automatically fill in the GM patch name for that
program for that track.

The rest of the options have been set up in a similar manner,
but do not actually have effect yet and are not recorded in
the song file or config file.