Added color tool. File format changed.
When color tool is activated you can edit the colors of blocks.
Drag left mouse on a block to change its color.
Middle mouse on a block will paste last color on it.
Right mouse on a block will paint the entire track the last color.

Clone blocks, that is, blocks which share the same pattern, will
share the same color. In order to make this arrangement sensible
I moved the color information from blocks to patterns and had to
change the file format.