Option record pass-thru now works.
record pass-thru: midi input will be echoed to midi output
on the port of currently selected track. You can now deactivate
this if it is annoying.

BUG: midi input is sent on frame zero of the current frame
buffer each time echo is processed. This may introduce error
on the order of nframes / samplerate = several milliseconds.
The midi input time must be converted to an appropriate frame
number on output to fix this.