New options follow playback and record on channel.
Follow playback: the screen will scroll to keep up with
the playback. This may be useful if you are playing along
with the song and need a visual cue. You can now deactivate
this if it is annoying.

Record on channel: if this is not activated, midi input
will be recorded on the currently selected track indicated
by the little keyboard on the left. If this is activated
midi will instead be recorded to the first track that
matches the channel of the messages. This may be useful
if you wish to control the target track by changing the
channel on your midi input device.