New option always hard copy.
Always hard copy: when you paste a block it will contain a
copy of the source block's pattern. If this is not activated
then copying will cause the new block to have the same pattern;
they will share the pattern. Any changes made to one block will
affect the other. You can now deactivate this 'cloning' behavior
if it is annoying.