Fixed timelines. For real this time. I think.
There were several bugs, some of which I did not
know about.

Selecting play position and loop positions from inside
pattern editor in a block not at time zero with zoom...
failed to set the positions correctly. That is fixed.

The play position pointer did not update correctly
when zoom changed or when you switched between blocks
while the transport is paused. That is fixed.

The root of these problems is simply that the formulas
for the timeline that I wrote are too complex to
understand. Also the two timelines, while they look the
same, are very different. However I chose to make only
one timeline class. This was a mistake and may be
fixed in the future.