GUI Changes. Added midi scope.
The midi scope displays all midi input in raw form and
annotated form. Currently it only works while the sequencer
is playing or recording.

Added the keyboard layout panel. It does not yet work.

Added a metronome widget. It does not yet work.

Added a menu in the configuration panel to decide what
to do when recording outside a block. The options do not
work yet (except for do nothing).