Added basic metronome. Fixed a bug. Scope works.
The metronome widget ticks when playing or recording.
The beats per measure affects the display of the
metronome. Better graphics would be nice here.

There was a bug in the song editor which failed to
place new blocks correctly when not using 4 beats per
measure. Fixed.

The monitor callback which updates the midi scope
now runs continuously. When not in playback mode,
the callback runs generally ten times a second. In
order to update widgets faster during playback it
automatically speeds up to 100 times per second and
returns to 10 after stopping.