Reduced monitor callback period. Webpage.
I was experiencing some jitter, I think, when watching
the metronome. It may have been imagined, but the previous
monitor period of 10ms may very well be noticably bad
for some people. The monitor now intends to run at 5ms
when playing or recording. The idle period is still
100ms.

A simple webpage is now in doc/index.html. The screenshots
are not in the repo and they will change anyway as I
get better graphics.