Keyboard customize works. Slightly new qwerty layout.
Now [ ] = and \ play notes F F# G and G# respectively.
Octave up, octave down, and zoom in were changed to
shift+] shift+[ and shift+= respectively.

You can now change the keyboard layout in the
configuration panel. Click on a keygrabber widget
then press the key or key combination. That will
set the desired action to that combination. The
names of the keys are probably english only, but
hopefully this lets non qwerty users arrange the
keyboard so that it is playable.

Saving and loading the keyboard layout to the
config file does not yet work.