LASH support added.
LASH is now a required dependency until I can
determine how to make it optional via autoconf. The
defined constant HAVE_LASH will activate or deactivate
LASH code when building jack.cpp, so all that would
be needed is to define or not define that depending
on if LASH is detected by the configure script.

After testing on LASH 5.4, I was not able to reload
epichord, lash_synth, or dino to their previous state
after saving. The LASH_Restore_File event is simply
not sent to epichord upon running. The following
error from lashd is the only clue.

conn_mgr_send_server_event_to_client: could not
send event to unknown connection id 18

where 18 increments each time this failure occurs.

Curiously non sequencer, when running alone, avoids
this fate most of the time, though it does fail to
restore non sometimes. It also fails to load non
most of the time when running with other programs.

When running lash programs together, it seems random
whether the restore will be sent. Most of the time
it is not, but sometimes it is.

I will need to determine if this is normal behavior
for lash 5.4 or it is just my system. It is, as far
as I can tell, not epichord.