Added unsafe block resizing. Broke restating.
You can now left click drag on block edges to
resize them. If you have many blocks selected
it will resize all of them.

There is not yet a guard against resizing so
as to overlap other blocks. Doing this may
cause a crash while playing.

There is not yet support for resizing to negative
width. This is the reason there is not yet a
test for resize safety. If some of the blocks are
resized to negative width, their resize command
will fail. Writing a collision test for this
behavior would be too much work if it is
temporary.

If you resize, move, or delete (and later paste)
blocks during an intersection with the play position,
the changes will not be immediately registered by
the sequencer. This is caused by a refactoring of
the 'restating' policy and a rewriting of the
restate function. The restate function is being
called at the right time, but is apparently not
working correctly.