Restate bugs fixed.
When a block is moved, resized, or deleted such that
the sequencer state immediately must change, it now
changes immediately.

Paste probably still behaves wrong but the main
cause of this bug has been fixed. Arranger paste
function does not use the new arranger framework
which was modeled after the later-written event
editor. Paste will be fixed when paste is rewritten
to be fully functional.