Updated documentation.
There is now a BUGS file which lists bugs as soon as
I remember them (and encounter new bugs).

Controls docs is more finalized.

The manual was rewritten, and is still incomplete. The
final version will take the form of a webpage with a
diagram.

The README now correctly lists where to get the source.