Added more pre-functional recording support.
The layerstack sequencer element stores an array of
patterns. Some blocks may contain this. I added a
flag so the sequencer knows if a block is to be
overwritten or layered when first recorded upon.

I added some reference counting stuff to patterns.
Since blocks, at this time, are never deleted (because
the undo invokations have not been written) it is
largely untested.