Heavy changes to file code. New format.
Semi-big change: The global list of patterns
was removed. This was a source of segfaults
and complicated midi import. The only reason
it existed was to make saving easier. It
was replaced with a function to collect and
find all unique patterns that are visible
in the arranger.

The file code and file format was changed. Now the
files have some magic header and other info that
should keep accidental opening of non song files
from crashing the program or putting it in an
inconsistent state.

The file stores layers and current layer now.

The file contains a file version field, which I hope
does not have to be used.

The file contains the ticks per beat that was used
by the build of epichord that saved it. This will be
used to convert to whatever the local build is using.
At some point the ticks per beat can be changed
in backend.h, but now there is a magic 128 all over.
After it can be changed easily, it will likely be
'fixed' at some number bigger than 128. This will
be used by most builds unless some people customize
their own copy. This addition to the file format will
make their songs portable.

Some gui changes were made. The menu options new,
load, and import now call a uihelper function to reset
the program state in case of error. Previously the
program was left in an inconsistent state. Extra windows
now appear if you press their button while they are
already opened. The fltk docs say that they should come
to the top anyway, but they don't. This will hopefully
be an acceptable work-around.

The file format assumed by this branch is not yet stable.
I wish to reinterpret the event times as delta time rather
than absolute time to save space.

Plans before this branch is to be merged include importing
midi files and finalizing the text file format (version 0).