SMF import minimally works, unsafely.
It will load and play the events. But the track
settings are left in an inconsistent state and
manipulating them may cause a segfault. There are
other less critical missing features.

The track names, tempo, and meta events have no
effect on the gui. All the created blocks are
the same color. The events are all there, but
the notes do not appear to be the correct
length because their durations were not updated.