Improved importing. Safer.
You can now edit patterns that were imported.
A bug was fixed that crashed if in the arranger
you drew a box above or below the valid tracks.

The track_info widget can now have more or less
than 16 modules. A module is added for each
track that is read from the midi file.