Added more importing features.
Note lengths appear correctly in the pattern editor.

Tempo changes applied to the song, but it applies
all of them in order at once. This leaves you with
the last tempo of the song. At this time it needs
to just take the first one.

There is a bug in the arranger. It currently tracks
the last block that was checked for resize handles.
When songs are loaded this pointer becomes invalid.
It has caused a crash. It needs to be fixed, somehow.

A bug when reading the microseconds per quarter note
was fixed.