Import track names added.
The track names in the midi file are used as the
track names for the tracks. The track structure
was changed so that instead of an array of characters
the name is a char pointer. The defined constant
MAX_TRACK_NAME was removed. There is no longer
a limit on track name length.