Fixed bug that crashed on saving after importing.
The track names are stored as allocated char buffers.
If no track name event appears for a track in a
midi file, it should store an empty string for that
track. Now it does that.

Side note, the transferring of strings between various
files and buffers is currently dont using strncpy.
This function has a bug about null termination and
also it will not allow embedded null bytes in a track
name. This needs to be addressed.