Initial refactoring of the scroll stuff.
The good news is that the graphical gliches are gone
and the layout kludge is obliterated. The bad news
is that you cannot scroll in the song editor.

The point of this branch is to replace both the
song_scroll and pattern_scroll fltk::ScrollGroups
with two scroll bars which simply control the widgets
using their callback.

The resulting sanity will make the graphical problems
trivially evaporate.