Song vertical scroll works.
It now acts like the old scroll bar from the scrollgroup.
I added a layout function to the arranger which serves
as the only way that widgets will get updates about the
window resizing. It required a kludge similar to the
scrollgroup. The fact that the magic kludge value is 2
rather than 4 means the current scheme is twice as good.