Enabled undo and redo. Buggy. Fixed key modifiers.
Undo and redo are activated with control z and
control r, and cannot be remapped yet.

It certainly will fail and eventually crash if
you fool with layers. I also got it to crash
when doing heavy editting in the piano roll.

Undoing recorded commands is currently klunky
and unusable.