Fixed two major sequencer dispatch bugs.
Events are now sorted in time before they are
dispatched. This solves rather large timing errors
in which the kludge was to send all out of order
events and following events on the last index of
the frame buffer.

An arithmetic error was causing ticks and target
frames for events to be off by 1, sometimes. This
mostly caused the above kludge to trigger, but sometimes
resulted in a negative target frame (zero minus one).
That would cause a segfault.

Fixed.