Added triplet quantization.
This was a much wanted feature. Toggle the triplet button in
the pattern editor and notes will snap to time positions which
are as close to thirds of grid vertical rules as possible. The
quantization resolution settings affect what exactly is being
divided in three.

At this time there is a bug. A beat is divided into 128 ticks which
is not divisible by three. This is semi hard coded and needs to be
made configurable at compile time. A default value of 288 would make
more sense.