Triplet resize works better now, sorta.
But really the q_tick variable needs to be changed to reflect
the triplet quantization. Actually that will not work if the
ticks cannot be divided by three. Once I change TICKS_PER_BEAT
this will not be a problem of eigth note or sixteenth note
mode, but not necessarily for others. Must investigate further.