Fixed problematic, wrong jack transport things.
An undue latency on looping was removed as well as incorrect
implementation of jack transport. Now play, stop, and reposition
will send jack transport messages, but epichord will not respond
to such messages itself. I have an idea of how this could be
correctly done, but it would take some work.