Dynamically updating controller widgets, and now in decimal!
When volume, pan, or program change on a track, the widgets in
the track settings module update to show the change. They also
display the current value in decimal when you are editing them.
This makes the program widget on the second panel actually useful.

The bank does not update and needs more work before it can be
fully functional (MSB/LSB).

The pan widget now has a different display.