Fixed bug in resize.
It was possible that the resize operation would move the completely
wrong note off if resizing from the right. The work around was to
have the search for note off begin right after the note off instead
of at the same position. The effect is that it cannot find note on
off pairs that effectively give the note duration zero.